[REFAC] Move class LayerData to layerdata.c/h
