+++ src/libs/antares/study/load.cpp
@@ -456,47 +456,5 @@ bool Study::reloadXCastData()
     return ret;
 }
 
-void LayerData::loadLayers(const AnyString& filename)
-{
-    IniFile ini;
-    if (std::ifstream(filename.c_str()).good()) // check if file exists
-        if (ini.open(filename))
-        {
-            // The section
-            auto* section = ini.find("layers");
-            if (section)
-            {
-                size_t key;
-                CString<50, false> value;
-
-                for (auto* p = section->firstProperty; p; p = p->next)
-                {
-                    // We convert the key and the value into the lower case format,
-                    // since several tests will be done with these string */
-                    key = p->key.to<size_t>();
-                    value = p->value;
-
-                    layers[key] = value.to<std::string>();
-                }
-
-                section = ini.find("activeLayer");
-                if (section)
-                {
-                    auto* p = section->firstProperty;
-                    activeLayerID = p->value.to<size_t>();
-
-                    p = p->next;
-
-                    if (p)
-                        showAllLayer = p->value.to<bool>();
-                }
-                return;
-            }
-
-            logs.warning() << ": The section `layers` can not be found";
-            return;
-        }
-}
-
 } // namespace Data
 } // namespace Antares
