+++ src/libs/antares/study/save.cpp
@@ -266,36 +266,5 @@ bool Study::saveToFolder(const AnyString& newfolder)
     return ret;
 }
 
-bool LayerData::saveLayers(const AnyString& filename)
-{
-    IO::File::Stream file;
-    if (file.openRW(filename))
-    {
-        CString<256, true> data;
-        data << "[layers]\n";
-        for (std::map<size_t, std::string>::iterator iterator = layers.begin();
-             iterator != layers.end();
-             iterator++)
-        {
-            data << iterator->first << " = " << iterator->second;
-            data << '\n';
-        }
-
-        data << "[activeLayer]\n";
-        data << "activeLayerID"
-             << " = " << activeLayerID;
-        data << '\n';
-        data << "showAllLayer"
-             << " = " << showAllLayer;
-
-        file << data;
-
-        return true;
-    }
-
-    logs.error() << "I/O error: impossible to write " << filename;
-    return false;
-}
-
 } // namespace Data
 } // namespace Antares
